refactor(ndi): extract compat_controller (+ consumer_build) modules — consumers.rs under the 1000-line cap
Pure code movement to satisfy the CI file-size gate (consumers.rs was 1235
lines, hard cap is 1000). No behavior change.

- compat_controller.rs: CONTROLLER_INTERVAL, spawn_compat_bitrate_controller,
  LossSample (+Default/+delta), read_loss_sample, stats_i64, stats_u64.
- consumer_build.rs: build_consumer_elements (+ vp8enc/h264/vp8 payloader
  builders), add/link_consumer_elements, adopt_encoder_timeline,
  spawn_consumer_bus_watch, connect_branch_signals.

The orchestrator build_consumer_pipeline_blocking and all ConsumerBranch /
lifecycle / negotiation logic stay in consumers.rs unchanged.
